Moved CallHasInstrument to a separate file.
