+++ apps/backend/src/datasources/mockups/CallDataSource.ts
@@ -1,8 +1,5 @@
-import {
-  AllocationTimeUnits,
-  Call,
-  CallHasInstrument,
-} from '../../models/Call';
+import { AllocationTimeUnits, Call } from '../../models/Call';
+import { CallHasInstrument } from '../../models/CallHasInstrument';
 import { CreateCallInput } from '../../resolvers/mutations/CreateCallMutation';
 import {
   AssignInstrumentsToCallInput,
