+++ apps/backend/src/models/CallHasInstrument.ts
@@ -1,0 +1,9 @@
+export class CallHasInstrument {
+  constructor(
+    public callId: number,
+    public instrumentId: number,
+    public availabilityTime: number,
+    public submitted: boolean,
+    public sepId?: number
+  ) {}
+}
